feat(admin): split Features, API Keys, and Advanced into a dedicated Admin page
Move the admin-flavoured tabs out of /settings into a new /admin/:tab
surface, gated by canEditFeatureFlags. Legacy /settings/{features,
api-keys,advanced} URLs redirect to their new /admin home so existing
bookmarks keep working. Sidebar grows an Admin section that's only
visible to workspace admins.
